[11] - Eliminar Dockerfile y añadir imagen a docker-compose.yml
